refactor: relocate query objects to `Domain` layer and update imports
- Moved `GetAllTournamentsQuery`, `GetTournamentQuery`, and other query objects to `Domain` layer.
- Added `GetHistoryQuery` for handling username-based query logic.
- Updated import paths across affected files for consistency.
