+++ tournament/src/Domain/Queries/QueryObject/GetHistoryQuery.ts
@@ -1,0 +1,9 @@
+export class GetHistoryQuery
+{
+    public readonly username: string;
+
+    constructor(username: string)
+    {
+        this.username = username;
+    }
+}
